fix some error

+++ Agora-C_Sharp-RTC-SDK/Code/Rtm/IRtmPresence.cs
@@ -7,9 +7,9 @@ namespace Agora.Rtm
 
         public abstract int WhereNow(string userId, ref UInt64 requestId);
 
-        public abstract int SetState(string channelName, RTM_CHANNEL_TYPE channelType, StateItem[] items, UInt64 count, ref UInt64 requestId);
+        public abstract int SetState(string channelName, RTM_CHANNEL_TYPE channelType, StateItem[] items, int count, ref UInt64 requestId);
 
-        public abstract int RemoveState(string channelName, RTM_CHANNEL_TYPE channelType, string[] keys, UInt64 count, ref UInt64 requestId);
+        public abstract int RemoveState(string channelName, RTM_CHANNEL_TYPE channelType, string[] keys, int count, ref UInt64 requestId);
 
         public abstract int GetState(string channelName, RTM_CHANNEL_TYPE channelType, string userId, ref UInt64 requestId);
     }
